Multimodal Fusion/requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
antlr4-python3-runtime==4.13.1
anyio==4.2.0
apache-beam==2.48.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.0.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
audioread==3.0.1
autopep8==2.0.4
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.1.1
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.1
configparser==6.0.1
contourpy==1.2.0
cookiecutter==2.5.0
crcmod==1.7
cryptography==42.0.1
cycler==0.12.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.1.1
dlib==19.24.2
dm-tree==0.1.9
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring-to-markdown==0.13
docstring_parser==0.16
docutils==0.20.1
dotmap==1.3.30
easydict==1.13
einops==0.8.1
et_xmlfile==2.0.0
etils==1.12.2
exceptiongroup==1.2.0
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.3
fastavro==1.10.0
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
flake8==6.1.0
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.6.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.26.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gql==0.2.0
graphql-core==3.2.6
grpcio==1.65.5
h5py==3.10.0
hdfs==2.7.3
httplib2==0.22.0
huggingface-hub==0.29.1
idna==3.6
imagesize==1.4.1
immutabledict==4.2.1
importlib-metadata==7.0.1
importlib_resources==6.5.2
inflection==0.5.1
intervaltree==3.1.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.3
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
isort==5.13.2
jaraco.classes==3.3.0
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpickle==3.4.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.2
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==24.3.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
lxml==5.1.0
mapbox-earcut==1.0.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mne==1.10.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.9
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.0.7
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
numpydoc==1.6.0
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
objsize==0.6.1
opencv-python==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.15
overrides==7.6.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
patsy==1.0.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.1.0
pluggy==1.4.0
pooch==1.8.0
portalocker==3.1.1
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.26.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pydot==1.4.2
pydub==0.25.1
pyflakes==3.1.0
pygame==2.5.2
Pygments==2.17.2
pylint==3.0.3
pylint-venv==3.0.3
pyls-spyder==0.4.0
pymongo==4.11.1
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
python-slugify==8.0.2
pytoolconfig==1.3.1
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.3.0
qtconsole==5.5.1
QtPy==2.4.1
redis==5.2.1
referencing==0.32.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
retina-face==0.0.14
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rope==1.12.0
rpds-py==0.17.1
rsa==4.9
Rtree==1.2.0
safetensors==0.4.2
scenic==2.1.0
scikit-learn==1.4.0
scikit-posthocs==0.11.4
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==1.44.1
Shapely==1.8.5.post1
shortuuid==1.0.13
simple-parsing==0.1.7
six==1.16.0
smmap==5.0.1
-e git+https://github.com/vchoutas/smplx.git@1265df7ba545e8b00f72e7c557c766e15c71632f#egg=smplx
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
spyder==5.5.0
spyder-kernels==2.5.1
stack-data==0.6.3
statsmodels==0.14.5
subprocess32==3.5.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-datasets==4.9.8
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.17.1
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
textdistance==4.6.1
threadpoolctl==3.2.0
three-merge==0.1.1
timm==0.4.5
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
torch==1.12.1+cu113
torch-geometric==2.6.1
torchaudio==0.12.1+cu113
torchsummary==1.5.1
torchvision==0.13.1+cu113
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
transformers==4.38.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.11.0
tzdata==2023.4
ujson==5.9.0
uri-template==1.3.0
urllib3==2.1.0
wandb==0.8.32
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wget==3.2
whatthepatch==1.0.5
widgetsnbextension==4.0.10
wrapt==1.16.0
xlrd==2.0.2
XlsxWriter==3.1.9
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zstandard==0.23.0
